Lumen fixes
Workaround for DX11 AMD driver bug breaking world space radiance cache updates
Simple repro:
* Run default plane level on a AMD GPU with -dx11
* r.Lumen.RadianceCache.ForceFullUpdate 1
* r.Lumen.RadianceCache.Visualize 1
* Notice that some probes are black

Workaround for AMD driver bug breaking global SDF
* Enable Lumen or use r.AOGlobalDFResolution 224
* ShowFlag.VisualizeGlobalDistanceField 1
* Open default level with a plane, move plane away from the 1st clipmap, move it up and down, notice artifacts below it

Workaround for a bug in FMath::Clamp<int32>(int32, uint32, uint32) causing it to work as if it would be FMath::Clamp<uint32> and to overflow Lumen Scene atlas.

#lockdown Juan.Canada
#rb Danie.Wright
#preflight 6215174f840463c99a663425
#jira UE-143183

#ROBOMERGE-AUTHOR: krzysztof.narkowicz
#ROBOMERGE-SOURCE: CL 19075983 in //UE5/Release-5.0/... via CL 19094535
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v921-19075845)

[CL 19136652 by krzysztof narkowicz in ue5-main branch]

+++ Engine/Source/Runtime/Renderer/Private/Lumen/LumenScene.cpp
@@ -1030,8 +1030,8 @@ void FLumenCard::GetMipMapDesc(int32 ResLevel, FLumenMipMapDesc& Desc) const
 	check(ResLevel >= Lumen::MinResLevel && ResLevel <= Lumen::MaxResLevel);
 
 	const FIntPoint ResLevelBias = ResLevelToResLevelXYBias();
-	Desc.ResLevelX = FMath::Clamp<int32>(ResLevel - ResLevelBias.X, Lumen::MinResLevel, Lumen::MaxResLevel);
-	Desc.ResLevelY = FMath::Clamp<int32>(ResLevel - ResLevelBias.Y, Lumen::MinResLevel, Lumen::MaxResLevel);
+	Desc.ResLevelX = FMath::Clamp<int32>(ResLevel - ResLevelBias.X, (int32)Lumen::MinResLevel, (int32)Lumen::MaxResLevel);
+	Desc.ResLevelY = FMath::Clamp<int32>(ResLevel - ResLevelBias.Y, (int32)Lumen::MinResLevel, (int32)Lumen::MaxResLevel);
 
 	// Allocations which exceed a physical page are aligned to multiples of a virtual page to maximize atlas usage
 	if (Desc.ResLevelX > Lumen::SubAllocationResLevel || Desc.ResLevelY > Lumen::SubAllocationResLevel)
